🔐☕️↣ Adding contract metadata for [#18 #16 #6] and a rudimentary frontend component for connecting to APIs (including with Py) for #1
Two new directories:
`./Ansible` ↣ Contract & build files here. Will be moved into a new folder in the root React/Next.js project, but for now I'm leaving this in a separate node project to keep things separated for easier debugging

`./client` ↣ Frontend client for interacting with the smart contracts in `Ansible` directory. Similar story to the directory positioning in `Ansible`. We'll migrate it over to the Node project in the root directory before merging this branch, however we will also transition as much as possible out of typescript.