⛑️🐉 ↝ Adding YAML configuration for GraphQL interaction with #16
I've moved the main authentication function (Flask, Moralis) in the `server` directory into the `api/auth` dir. No longer appears to work (apparently an issue with `magic-sdk` not being included in another node package. We'll fix this later, but since we're getting Lens to handle the signature process for the next part of this webapp, that isn't going to be fixed for now.

https://github.com/signal-k/polygon/issues/24