🐤💸 ↝ Adding more profile contents & updating styling for Lens frontend (and defining some new roles for #16 backend)
Check out the Notion documentation here: https://skinetics.notion.site/Lens-posting-a07f0e9c243249c0a3517e4160874137

Currently having a bit of a problem with determining the method we'll use to define proposal parameters, but I've come up with a fallback solution (i.e. using bots & comments) for the MVP. Potentially these "comments" could also be stored off-Lens but still be visible on the frontend (problem here is if you're looking at the proposals from an alternate client), bringing the Discord integration further into the limelight (using DeWork.xyz as the intermediary bot), as well as highlighting a real usecase/benefit to having both Supabase & Moralis as the data stores. Hopefully the metadata standards for publications are just as loose as with the `appId` standard on lens, as we'll just be able to create dummy metadata that is saved to the IPFS publication URI and therefore callable via graphql.

So overall, a lot of questions coming up, but some UI improvements as well as some more read interactivity for the client...